Apply growth rates to initial_value datasets
This commit removes the ability to perform per-dataset aggregations
prior to combining datasets. That methodology did not work with growth
rate datasets. Secondly, while there are some efficiencies to be gained
with aggregating dimensions in that early stage, such as combining
electricity end uses, we expect that we will usually store cached
versions of project-mapped datasets. User queries will use those cached
datasets, and so this is not a significant loss in functionality.