Add null rows for expected-missing data
In the existing two-table format the load_data_lookup table designates
expected-missing dimension records by specifying id=null. This
information was getting dropped when the dataset was mapped to the
project. This commit adds that information to the mapped dataset by
adding a row where the timestamp and all value columns are null.