Disable Hive storage
The current code is causing errors on Kestrel with Spark 3.4.1. Whenever
we restart the Spark session, the new session reports that the Derby
database was booted by someone else and fails. The root cause is
currently unclear. I disabled Hive support and that prevents the issue
from happening. Caching of dimension record tables is still working, and
so we do not appear to need Hive tables any longer.