Hide tracebacks in CLI cmds for dsgrid exceptions
This commit changes the behavior of dsgrid exceptions that get raised in
CLI commands. Instead of letting the entire traceback get dumped to the
console, the code only displays summary information. This will make it
easier for users to figure out what failed.

There is an option to preserve the current behavior and reraise
exceptions.