Remove support for supplemental_columns in queries
The same functionality exists if the user specifies an AggregationModel
in the query. That is more complicated for the user to enter in cases
where they only want to add a column (and not perform an aggregation).

We made this decision for these reasons:
1. Simplify the code.
2. Provide a single way to perform this operation and simplify the user
   documentation.

If there is demand for this feature in the future, we could bring it
back but with a different implementation. Rather than perform an
independent aggregation (one for supplemental_columns, one for
AggregationModels), we could implicitly add the supplemental_columns
to all AggregationModels.