WIP: Add preliminary support for DuckDB
The goal of this work is to allow DuckDB to be used as the DataFrame
library instead of Spark. DuckDB is significantly faster for datasets
that can be processed on a single node, and especially for test & dev.

DuckDB offers experimental support for a subset of the pyspark API.
To complete this work we will likely need to rewrite pyspark queries
with direct SQL queries instead of the DataFrame API.